requirements.txt
anyio==4.4.0
audible==0.8.2
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
ffmpeg-python==0.2.0
future==1.0.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.8
pbkdf2==1.3
pillow==10.4.0
prompt_toolkit==3.0.47
pyaes==1.6.1
pyasn1==0.6.0
PyAudio==0.2.14
pydub==0.25.1
python-dotenv==1.0.1
requests==2.32.3
rfc3986==1.5.0
rsa==4.9
sniffio==1.3.1
soupsieve==2.6
tqdm==4.66.5
urllib3==2.2.2
Wave==0.0.2
wcwidth==0.2.13
